fix bug in pixel font cartouche middle
correct behavior: the left half of the cartouche, the middle of the cartouche, and the right half of the cartouche should all sample the same column of pixels to render, to ensure a smooth overlap.

previous, incorrect behavior: the left and right of the cartouche would correctly sample the same column, but the middle of the cartouche incorrectly sampled a column 1px to the left.

off-by-one error! :D