sheet v4: 8px fonts aren't blurry anymore
In sheet v4, and in all pixel fonts retroactively, the ascent and descent have changed.

Old: ascent 800, descent -200
New: ascent 875, descent -125

This means that for 8px fonts, the baseline rests at a pixel boundary: 1px is below the baseline, and 7px are above the baseline. So they won't be vertically blurry.

This also means that 8px fonts (that is, 7px of glyph, and 1px empty row at the bottom) rest on the baseline now, which isn't intentional but is neat.