vector-rotated critters
we rotate ni and critters based on whether the glyph has a `rotate` value in `default.toml`. we do the rotation using the vector data, so the diagonals deviate from the pixel grid instead of just looking broken.

we've removed some hard-coding, but there's still some left in ligature generation in `svgtottf.py`.

pixel font ni and critters become offset when rotated, because we don't take the assumed 1px margin into account when rotating.